move some view files around
