+++ packages/client/src/views/decorator-utils.ts
@@ -1,7 +1,7 @@
-import { defineCustomElement } from "./custom-elements";
 import { CustomEvents } from "../events";
 import { param } from "../params";
 import cls from "../styles/decorator-utils.module.css";
+import { defineCustomElement } from "./custom-elements";
 
 class DecoratorUtils extends HTMLElement {
     update = () => {
