+++ packages/client/src/views/dropdown-menu.ts
@@ -1,5 +1,5 @@
-import cls from "../styles/dropdown-menu.module.css";
 import { createEvent } from "../events";
+import cls from "../styles/dropdown-menu.module.css";
 import { defineCustomElement } from "./custom-elements";
 
 class DropdownMenu extends HTMLElement {
