+++ packages/client/src/views/screensharing-modal.ts
@@ -1,8 +1,8 @@
 import Cookies from "js-cookie";
-import { defineCustomElement } from "./custom-elements";
 import loadExternalScript from "../load-external-script";
 import { env, param } from "../params";
 import clsInputs from "../styles/inputs.module.css";
+import { defineCustomElement } from "./custom-elements";
 
 let hasBeenOpened = false;
